+++ R/native_cmd.R
@@ -13,9 +13,20 @@ native_cmd <- function(conda_cmd,
                        conda_args = NULL,
                        ...,
                        verbose = FALSE,
-                       stdout = "|") {
+                       error = c("cancel", "continue"),
+                       stdout = "|",
+                       stderr = "|") {
+  rlang::check_required(conda_cmd)
+  error <- rlang::arg_match(error)
+  if (isTRUE(identical(error, "cancel"))) {
+    error_var <- TRUE
+  } else {
+    error_var <- FALSE
+  }
+
   umamba_bin_path <- micromamba_bin_path()
   env_root_dir <- get_install_dir()
+
   if (isFALSE(fs::file_exists(umamba_bin_path))) {
     install_micromamba(force = TRUE)
   }
@@ -34,15 +45,30 @@ native_cmd <- function(conda_cmd,
       CONDA_PROMPT_MODIFIER = "",
       MAMBA_PROMPT_MODIFIER = "",
       CONDA_DEFAULT_ENV = "",
-      MAMBA_DEFAULT_ENV = ""
+      MAMBA_DEFAULT_ENV = "",
+      R_HOME = ""
     )
   )
+
+  if (isFALSE(stderr %in% c("|", ""))) {
+    verbose <- FALSE
+  }
+  callback_fun_out <- NULL
+  callback_fun_err <- NULL
+  # if (isTRUE(verbose)) {
+  #   callback_fun_out <- function(x, y) {
+  #     cli::cli_inform(x)
+  #   }
+  #   callback_fun_err <- function(x, y) {
+  #     x <- stringr::str_squish(x)
+  #     cli::cli_alert("{.red {x}}")
+  #   }
+  # }
   px_res <- processx::run(
     command = fs::path_real(umamba_bin_path),
     args = c(
       "--no-rc",
       "--no-env",
-      # "--log-level", "3",
       conda_cmd,
       "-r",
       env_root_dir,
@@ -53,7 +79,10 @@ native_cmd <- function(conda_cmd,
     echo_cmd = verbose,
     echo = verbose,
     stdout = stdout,
-    error_on_status = TRUE
+    stdout_line_callback = callback_fun_out,
+    stderr = stderr,
+    stderr_line_callback = callback_fun_err,
+    error_on_status = error_var
   )
   return(invisible(px_res))
 }